+++ src/main/java/com/sistema/examenes/services/ObjetivoODS_ServiceImpl.java
@@ -8,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -34,7 +33,7 @@ public class ObjetivoODS_ServiceImpl extends GenericServiceImpl<ObjetivoODS, Lon
 
         for (Object[] resultado : resultados) {
             ObjetivoOds_DTO objetivoDTO = new ObjetivoOds_DTO(
-                    ((BigInteger) resultado[0]).longValue(),
+                    (Long) resultado[0],
                     (String) resultado[1],
                     (String) resultado[2]
             );
